requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
beautifulsoup4 #used to scrap data
requests
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
groq
langchain-groq
langchain-core
langchain-classic
streamlit